--- src/App.css
@@ -1,34 +1,0 @@
-.App {
-  display: flex;
-  flex-direction: column;
-  margin-top: 0;
-  height: 100%;
-  overflow-x: hidden;
-}
-.sidebar {
-  overflow-y: scroll;
-  margin: auto;
-  overflow-x: hidden;
-}
-
-
-.map {
-  width: 100%;
-  height: 100%;
-}
-
-@media only screen and (min-width: 768px) {
-  .App {
-    flex-direction: row;
-  }
-  .sidebar {
-    height: 100%;
-    max-height: 100vh;
-  }
-
-  .map {
-    width: 80%;
-    height: 100vh;
-    max-height: none;
-  }
-}
